fix: address review feedback — extract shared EdgePanel, document gravity
- Extract duplicated _EdgePanel to shared EdgePanel widget in edge_panel.dart
- Add doc comment on gravity default explaining FR force scaling rationale
