+++ lib/src/ui/graph/force_directed_graph_widget.dart
@@ -7,6 +7,7 @@ import '../../engine/mastery_state.dart';
 import '../../models/knowledge_graph.dart';
 import '../../models/network_health.dart';
 import 'catastrophe_painter.dart';
+import 'edge_panel.dart';
 import 'graph_edge.dart';
 import 'graph_node.dart';
 import 'graph_painter.dart';
@@ -410,7 +411,7 @@ class _ForceDirectedGraphWidgetState extends State<ForceDirectedGraphWidget>
         child: Material(
           elevation: 0,
           color: Colors.transparent,
-          child: _EdgePanel(edge: edge),
+          child: EdgePanel(edge: edge),
         ),
       ),
     );
@@ -668,76 +669,3 @@ class _TeamNodePanel extends StatelessWidget {
   }
 }
 
-/// Tap card showing relationship details.
-class _EdgePanel extends StatelessWidget {
-  const _EdgePanel({required this.edge});
-
-  final GraphEdge edge;
-
-  @override
-  Widget build(BuildContext context) {
-    final color = edge.isDependency
-        ? Colors.white.withValues(alpha: 0.9)
-        : Colors.white.withValues(alpha: 0.6);
-
-    return Card(
-      elevation: 4,
-      child: Container(
-        constraints: const BoxConstraints(maxWidth: 250),
-        padding: const EdgeInsets.all(12),
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Row(
-              mainAxisSize: MainAxisSize.min,
-              children: [
-                Icon(
-                  edge.isDependency ? Icons.arrow_forward : Icons.link,
-                  size: 14,
-                  color: color,
-                ),
-                const SizedBox(width: 6),
-                Flexible(
-                  child: Text(
-                    edge.label,
-                    style: TextStyle(
-                      fontWeight: FontWeight.bold,
-                      fontSize: 13,
-                      color: color,
-                    ),
-                  ),
-                ),
-              ],
-            ),
-            const SizedBox(height: 8),
-            Text(
-              '${edge.source.name}  \u2192  ${edge.target.name}',
-              style: const TextStyle(fontSize: 12),
-            ),
-            if (edge.relationship.description != null) ...[
-              const SizedBox(height: 4),
-              Text(
-                edge.relationship.description!,
-                style: const TextStyle(fontSize: 11, color: Colors.grey),
-                maxLines: 3,
-                overflow: TextOverflow.ellipsis,
-              ),
-            ],
-            if (edge.isDependency) ...[
-              const SizedBox(height: 4),
-              Text(
-                'Dependency',
-                style: TextStyle(
-                  fontSize: 10,
-                  fontWeight: FontWeight.w600,
-                  color: Colors.amber.shade300,
-                ),
-              ),
-            ],
-          ],
-        ),
-      ),
-    );
-  }
-}
